+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* `MainThreadMarker::new_unchecked` and `MainThreadBound::new` is now available
+  in `const`. This is useful for creating main-thread only statics.
+
 
 ## 0.2.2 - 2024-05-21