+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -13,8 +13,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased - YYYY-MM-DD
 
 ### Added
-* `MainThreadMarker::new_unchecked` and `MainThreadBound::new` is now available
-  in `const`. This is useful for creating main-thread only statics.
+* `objc2-foundation`: `MainThreadMarker::new_unchecked` and
+  `MainThreadBound::new` is now available in `const`. This is useful for
+  creating main-thread only statics.
+* `objc2-foundation`: `MainThreadMarker::from` now debug-asserts that it is
+  actually running on the main thread.
 
 
 ## 0.2.2 - 2024-05-21