+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -27,6 +27,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   prevents real-world usage of these types, and isn't actually needed for
   soundness (the documentation mentions the collection being "corrupt" if the
   hash is changed, but that's not the same as it being unsound).
+* Updated SDK from Xcode 15.4 to 16.0.
+
+  View the release notes to learn more details:
+  - [16.0](https://developer.apple.com/documentation/xcode-release-notes/xcode-16-release-notes)
 
 ### Deprecated
 * `objc2-foundation`: Moved `MainThreadMarker` to `objc2`.